AK3: Fix patch_fstab function
Patching system partition entry won't work if system_ext partition is present. Fix this by adding [[:space:]] to the end of the pattern.
Note that grep -w can't be used in this case because there is still a possibility of erroneous matches, e.g., logical vendor partition (that is just "vendor") and "/vendor/dsp".

Reported-by: Wiley Lau <henloboii@protonmail.ch>